app/requirements.txt
streamlit
langgraph
python-dotenv
langchain-huggingface
langchain-core
langchain-text-splitters
langchain-community
langchain-pinecone
langchain-classic
langchain-cohere
pinecone
langchain-groq